test(static-verifier): keep constraints-only prover helper private
